openjdk integration
git-svn-id: svn+ssh://svn.code.sf.net/p/jnode/svn/trunk@4709 407389f7-7c16-0410-9eee-ad85acb55d7f

+++ core/src/openjdk/java/java/lang/StringBuffer.java
@@ -76,7 +76,6 @@ package java.lang;
  * it performs no synchronization.
  *
  * @author	Arthur van Hoff
- * @version 	1.107, 05/05/07
  * @see     java.lang.StringBuilder
  * @see     java.lang.String
  * @since   JDK1.0
@@ -642,6 +641,6 @@ package java.lang;
         throws java.io.IOException, ClassNotFoundException {
         java.io.ObjectInputStream.GetField fields = s.readFields();
         value = (char[])fields.get("value", null);
-        count = (int)fields.get("count", 0);
+        count = fields.get("count", 0);
     }
 }
